tools: Fix update-protocol-ids to parse enum items
Augment "update-protocol-ids" to correctly parse newer versions of the
"ndpi_protocol_ids.h" header which use an enum to define the values of
the protocol identifiers. To choose the parsing mode for new style
headers the "--enum" command line option needs to be passed, while the
parsing of old style headers is chosen with "--defs". A new "--help"
(or "-h") option is recognized as well.